back/app/requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-extensions==2.1.9
django-filter==2.1.0
django-mysql==3.2.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
dnspython==1.16.0
eventlet==0.25.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
Markdown==3.1.1
monotonic==1.5
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
